+++ test/ch/stream_test.exs
@@ -48,14 +48,7 @@ defmodule Ch.StreamTest do
                  |> Stream.chunk_every(100_000)
                  |> Stream.map(fn chunk -> RowBinary.encode_rows(chunk, _types = ["UInt64"]) end)
                  |> Stream.take(10)
-                 |> Enum.into(
-                   Ch.stream(
-                     conn,
-                     "insert into collect_stream(i) format RowBinary",
-                     _params = [],
-                     encode: false
-                   )
-                 )
+                 |> Enum.into(Ch.stream(conn, "insert into collect_stream(i) format RowBinary\n"))
                end)
 
       assert Ch.query!(conn, "select count(*) from collect_stream").rows == [[1_000_000]]